完成使用切面Aspect实现公共字段填充功能

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -2,25 +2,33 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.annotation.AutoFill;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
-import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.enumeration.OperationType;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetMealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishMapper dishMapper;
+    @Autowired
+    private SetMealMapper setMealMapper;
 
     /**
      * 新增分类
@@ -28,14 +36,15 @@ public class CategoryServiceImpl implements CategoryService {
      * @param categoryDTO 添加分类的数据传输对象
      */
     @Override
+    @AutoFill(OperationType.INSERT)
     public void insert(CategoryDTO categoryDTO) {
         Category category = new Category();
         BeanUtils.copyProperties(categoryDTO, category);
         category.setStatus(StatusConstant.DISABLE);     // 默认新增的分类状态为禁用
-        category.setCreateTime(LocalDateTime.now());
+        /*category.setCreateTime(LocalDateTime.now());
         category.setUpdateTime(LocalDateTime.now());
         category.setCreateUser(BaseContext.getCurrentId());
-        category.setUpdateUser(BaseContext.getCurrentId());
+        category.setUpdateUser(BaseContext.getCurrentId());*/
         categoryMapper.insert(category);
     }
 
@@ -60,12 +69,13 @@ public class CategoryServiceImpl implements CategoryService {
      * @param status 需要修改的分类目标状态
      */
     @Override
+    @AutoFill(OperationType.UPDATE)
     public void updateStatus(Long id, Integer status) {
         Category category = Category.builder()
                 .id(id)
                 .status(status)
-                .updateTime(LocalDateTime.now())
-                .updateUser(BaseContext.getCurrentId())
+                /*.updateTime(LocalDateTime.now())
+                .updateUser(BaseContext.getCurrentId())*/
                 .build();
         categoryMapper.update(category);
     }
@@ -76,11 +86,12 @@ public class CategoryServiceImpl implements CategoryService {
      * @param categoryDTO 修改分类的数据传输对象
      */
     @Override
+    @AutoFill(OperationType.UPDATE)
     public void update(CategoryDTO categoryDTO) {
         Category category = new Category();
         BeanUtils.copyProperties(categoryDTO, category);
-        category.setUpdateTime(LocalDateTime.now());
-        category.setUpdateUser(BaseContext.getCurrentId());
+        /*category.setUpdateTime(LocalDateTime.now());
+        category.setUpdateUser(BaseContext.getCurrentId());*/
         categoryMapper.update(category);
     }
 
@@ -91,6 +102,17 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void delete(Long id) {
+        // 检查菜品分类中有无关联菜品
+        Integer dishCount = dishMapper.getCount(id);
+        if (dishCount > 0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+        // 检查套餐分类中有无关联套餐
+        Integer setMealCount = setMealMapper.getCount(id);
+        if (setMealCount > 0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
+        // 都没有关联则执行删除
         categoryMapper.delete(id);
     }
 }
